cleanup: a fixture level less, comment outdated since #106
