+++ src/test/scala/nl.knaw.dans.easy.deposit/servlets/ServletFixture.scala
@@ -19,33 +19,36 @@ import java.io
 import java.util.Base64
 
 import better.files.File
-import nl.knaw.dans.easy.deposit.EasyDepositApiApp
 import nl.knaw.dans.easy.deposit.authentication.AuthenticationProvider
+import nl.knaw.dans.easy.deposit.docs.DepositInfo
+import nl.knaw.dans.easy.deposit.{ EasyDepositApiApp, TestSupportFixture }
+import nl.knaw.dans.lib.error._
+import org.eclipse.jetty.http.HttpStatus._
 import org.scalatra.test.EmbeddedJettyContainer
 import org.scalatra.test.scalatest.ScalatraSuite
 
-/**
- * This Suite relies on Jetty 9.x, while we still require Jetty 8.x
- * By overriding localPort and baseUrl below, issues related to these versions are solved.
- */
-trait ServletFixture extends EmbeddedJettyContainer {
+trait ServletFixture extends TestSupportFixture with EmbeddedJettyContainer with ScalatraSuite {
   this: ScalatraSuite =>
 
   def mountServlets(app: EasyDepositApiApp, authenticationProvider: AuthenticationProvider): Unit = {
     val userServlet = new UserServlet(app) with UndoMasking {
       override def getAuthenticationProvider: AuthenticationProvider = authenticationProvider
     }
-    val depositServlet = new DepositServlet(app) with UndoMasking {
-      override def getAuthenticationProvider: AuthenticationProvider = authenticationProvider
-    }
     val authServlet = new AuthServlet(app) with UndoMasking {
       override def getAuthenticationProvider: AuthenticationProvider = authenticationProvider
     }
     val appServlet = new EasyDepositApiServlet(app) with UndoMasking
-    addServlet(depositServlet, "/deposit/*")
+    mountDepositServlet(app, authenticationProvider)
     addServlet(userServlet, "/user/*")
     addServlet(authServlet, "/auth/*")
     addServlet(appServlet, "/*")
+  }
+
+  def mountDepositServlet(app: EasyDepositApiApp, authenticationProvider: AuthenticationProvider): Unit = {
+    val depositServlet = new DepositServlet(app) with UndoMasking {
+      override def getAuthenticationProvider: AuthenticationProvider = authenticationProvider
+    }
+    addServlet(depositServlet, "/deposit/*")
   }
 
   /**
@@ -68,5 +71,18 @@ trait ServletFixture extends EmbeddedJettyContainer {
   val fooBarBasicAuthHeader: (String, String) = {
     val encoded = Base64.getEncoder.encodeToString("foo:bar".getBytes())
     ("Authorization", s"Basic $encoded")
+  }
+
+  /** @return UUID of the created deposit */
+  def createDeposit: String = {
+    post("/deposit/", headers = Seq(fooBarBasicAuthHeader)) {
+      // prevent json to interpret a stack trace
+      status shouldBe CREATED_201
+
+      // return the new UUID
+      DepositInfo(body)
+        .map(_.id.toString)
+        .getOrRecover(e => fail(e.toString, e))
+    }
   }
 }
